Fixed "empty regexp" issue; added regexp syntax checking; added error badge
1. The extension didn't behave properly when regexps were empty (due to
my shitty testing). Fixed.
2. When user provides regexp with incorrect syntax, we catch this and
show a descriptive error in extension's popup.
3. Error badge introduced. Use case: when user causes an error and
ignores it, our error badge will remind him that he need to fix it.
It's important because ignoring this extension may often result in not
being aware that Icinga/Nagios reports errors that we need to take care
of.